Refactoring eig_tools
